Price List - Class migration
