+++ classes/class-uagb-block-helper.php
@@ -2488,10 +2488,11 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 			);
 
 			// @codingStandardsIgnoreEnd
+			$base_selector = ( $attr['classMigrate'] ) ? '.uagb-block-' : '#uagb-rm-';
 
-			$desktop = UAGB_Helper::generate_css( $selectors, '#uagb-rm-' . $id );
-			$tablet  = UAGB_Helper::generate_css( $t_selectors, '#uagb-rm-' . $id );
-			$mobile  = UAGB_Helper::generate_css( $m_selectors, '#uagb-rm-' . $id );
+			$desktop = UAGB_Helper::generate_css( $selectors, $base_selector . $id );
+			$tablet  = UAGB_Helper::generate_css( $t_selectors, $base_selector . $id );
+			$mobile  = UAGB_Helper::generate_css( $m_selectors, $base_selector . $id );
 
 			$generated_css = array(
 				'desktop' => $desktop,
